[API-7676]: Validate apiKey is not null when creating a Client
